convert to big decimal

+++ app/src/main/java/com/shyam/currencyconverter/data/source/local/database/entities/CurrencyRates.kt
@@ -2,6 +2,7 @@ package com.shyam.currencyconverter.data.source.local.database.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity(tableName = "CurrencyRates")
-data class CurrencyRates(@PrimaryKey val base: String, val timestamp: Long, val rates: Map<String, Double>)
+data class CurrencyRates(@PrimaryKey val base: String, val timestamp: Long, val rates: Map<String, BigDecimal>)
