+++ app/src/main/java/com/shyam/currencyconverter/data/source/remote/network/response/GetCurrencyRatesResponse.kt
@@ -1,8 +1,9 @@
 package com.shyam.currencyconverter.data.source.remote.network.response
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 class GetCurrencyRatesResponse {
     @SerializedName("quotes")
-    var currencies: Map<String,Double>?=null
+    var currencies: Map<String,BigDecimal>?=null
 }
